+++ test-page/index.html
@@ -22,7 +22,7 @@ redirect_from:
     <body>
       <div id = 'title'>forest-plot</div>
       <div id = 'subtitle'>Example 1</div>
-      <div id= 'info'>This is a demo page for the interactive forest-plot tool created by the Rho. Code, details and documentation are available on <a href="https://github.com/jwildfire/forest-plot">github</a>.</div>
+      <div id= 'info'>This is a demo page for the interactive forest-plot tool, details and documentation are available on <a href="https://github.com/jwildfire/forest-plot">github</a>.</div>
       <div id = 'container'></div>
     </body>
